refact: #99 use store not to make duplicated getFriendList Api calls
