+++ src/stores/useBoundStore.ts
@@ -3,15 +3,15 @@ import { devtools } from 'zustand/middleware';
 import { immer } from 'zustand/middleware/immer';
 import { createMomentSlice, MomentSlice } from './moment';
 import { createMyPageSlice, MyPageSlice } from './my';
-import { createMyProfileSlice, MyProfileSlice } from './myProfile';
 import { createSignUpInfoSlice, SignUpInfoSlice } from './signUpInfo';
 import { createTodaysQuestionsSlice, TodaysQuestionsSlice } from './todaysQuestions';
+import { createUserSlice, UserSlice } from './user';
 
 export type BoundState = MomentSlice &
   MyPageSlice &
   SignUpInfoSlice &
   TodaysQuestionsSlice &
-  MyProfileSlice;
+  UserSlice;
 
 export type SliceStateCreator<Slice> = StateCreator<
   BoundState,
@@ -34,6 +34,6 @@ export const useBoundStore = create<BoundState>()(
     ...createMomentSlice(...a),
     ...createSignUpInfoSlice(...a),
     ...createTodaysQuestionsSlice(...a),
-    ...createMyProfileSlice(...a),
+    ...createUserSlice(...a),
   })),
 );
